Fix problems with images

+++ MilaIrisWebParent/MilaIrisBackEnd/src/main/java/kz/milairis/admin/category/CategoryController.java
@@ -40,7 +40,7 @@ public class CategoryController {
 		CategoryPageInfo pageInfo = new CategoryPageInfo();
 		List<Category> listCategories = service.listByPage(pageInfo, pageNum, sortDir, keyword);
 
-		long startCount = (long) (pageNum - 1) * CategoryService.ROOT_CATEGORIES_PER_PAGE + 1;
+		long startCount = (pageNum - 1) * CategoryService.ROOT_CATEGORIES_PER_PAGE + 1;
 		long endCount = startCount + CategoryService.ROOT_CATEGORIES_PER_PAGE - 1;
 		if (endCount > pageInfo.getTotalElements()) {
 			endCount = pageInfo.getTotalElements();
@@ -149,4 +149,3 @@ public class CategoryController {
 		exporter.export(listCategories, response);
 	}
 }
-
